more refactoring for module sake
