+++ src/components/Home/Home.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useDispatch } from 'react-redux';
 import { logout } from '../../store/actions';
 
-const Home = () => {
+export const Home = () => {
     const dispatch = useDispatch();
     return (
         <>
@@ -18,5 +18,3 @@ const Home = () => {
         </>
     );
 };
-
-export default Home;
