+++ src/components/shared/LoadingPage.tsx
@@ -11,7 +11,7 @@ const useStyles = makeStyles((theme) => ({
     },
 }));
 
-export default function LoadingPage() {
+export const LoadingPage = () => {
     const classes = useStyles();
 
     return (
@@ -19,35 +19,4 @@ export default function LoadingPage() {
             <CircularProgress />
         </div>
     );
-}
-
-// import React from 'react';
-// import { makeStyles } from '@material-ui/core/styles';
-// import CircularProgress from '@material-ui/core/CircularProgress';
-// import Backdrop from '@material-ui/core/Backdrop';
-
-// const useStyles = makeStyles((theme) => ({
-//     backdrop: {
-//         zIndex: theme.zIndex.drawer + 1,
-//         color: '#fff',
-//     },
-// }));
-
-// export default function LoadingPage() {
-//     const classes = useStyles();
-//     const [open, setOpen] = React.useState(false);
-//     const handleClose = () => {
-//         setOpen(false);
-//     };
-//     const handleToggle = () => {
-//         setOpen(!open);
-//     };
-
-//     return (
-//         <>
-//             <Backdrop className={classes.backdrop} open={open} onClick={handleClose}>
-//                 <CircularProgress color="inherit" />
-//             </Backdrop>
-//         </>
-//     );
-// }
+};
