+++ src/components/shared/index.ts
@@ -1,0 +1,5 @@
+export * from './Alert';
+export * from './Copyright';
+export * from './LoadingOverlay';
+export * from './LoadingPage';
+export * from './PrivateRoute';
