+++ src/containers/Main/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { BrowserRouter as Router, Route, Switch, Redirect } from 'react-router-dom';
-import Home from '../../components/Home/Home';
+import { Home } from '../../components/Home';
 
 const Main = () => {
     return (
